+++ api/src/google/mod.rs
@@ -7,7 +7,7 @@ pub use talaria_google_connections as connections;
 pub mod drive;
 pub use talaria_google_errors as errors;
 pub use talaria_google_gmail as gmail;
-pub mod oauth;
+pub use talaria_google_oauth as oauth;
 pub use talaria_google_org as org;
 pub mod pending_actions;
 pub use talaria_google_provisioning as provisioning;